Narrow TradeID to uint32_t
Matches OrderID, which is already uint32_t; 64-bit range was unnecessary headroom for a trade counter.

+++ cpp/include/models/trade.hpp
@@ -2,8 +2,7 @@
 #include <cstdint>
 #include "models/order.hpp"
 
-// TODO: make TradeID uint32_t
-using TradeID = uint64_t;
+using TradeID = uint32_t;
 
 class Trade {
 private:
